% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_diff.R
\name{compute_diff}
\alias{compute_diff}
\title{Calculation of differential methylation values for DMRs}
\usage{
compute_diff(
  DMR,
  beta,
  human_ref = c("Cancer", "Tissue"),
  Case_name,
  Control_name
)
}
\arguments{
\item{DMR}{the converted DMR that converted by function preinput}

\item{beta}{the beta matrix with case group and control group(The case group is in the front, the group group is in the back)}

\item{human_ref}{Human genome reference strand from different cancer tissue or different tissue}

\item{Case_name}{the colnames of case group}

\item{Control_name}{the colnames of control group}
}
\description{
This function calculates the differential methylation value of the DMR and the number of probes and CpGs included and the length of the DMR
}
\examples{
beta_file = system.file("extdata","example_beta.RDS",package = 'DMRArrayMetric')
example_beta = readRDS(beta_file)
case = colnames(example_beta)[1:5]
control = colnames(example_beta)[6:10]
res_DMR = compute_diff(DMR = convert_DMR,beta = example_beta, human_ref = 'Cancer', Case_name=case, Control_name=control)

}
