% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metric.R
\name{compute_metric}
\alias{compute_metric}
\title{Calculation metric DMRn for each DMRs}
\usage{
compute_metric(DMR)
}
\arguments{
\item{DMR}{the DMR is obtained by compute_diff()}
}
\description{
This function calculates the metric DMRn
}
\examples{
DMRn = compute_metric(DMR = res_DMR)

}
