% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preinput.R
\name{preinput}
\alias{preinput}
\title{Generate DMRs data that can be used for metric calculations}
\usage{
preinput(DMR, chr, start, end, minProbeLength = 3)
}
\arguments{
\item{DMR}{the raw DMR that need convert}

\item{chr}{the column index of chromosome in the raw DMR and The format should be character type 1,2,3...,X,Y}

\item{start}{the column index of start in the raw DMR}

\item{end}{the column index of end in the raw DMR}

\item{minProbeLength}{The minimum number of probes the DMR should have}
}
\description{
This function generatea DMR data that includes chromosome, starting and end position, DMR length, and probe distribution
}
\examples{
DMR_file = system.file("extdata","raw_DMR.csv",package = 'DMRArrayMetric')
DMR <- read.csv(DMR_file)
DMR$seqnames = as.character(DMR$seqnames)
convert_DMR <- preinput(DMR, chr=3, start=4, end=5, minProbeLength=3)
}
