% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_graph.R
\name{rank_graph}
\alias{rank_graph}
\title{Draw a line chart for DMRn.}
\usage{
rank_graph(DMRfilePath, threshold = seq(0, 0.9, 0.1), Title = NULL)
}
\arguments{
\item{DMRfilePath}{The text file containing the metric DMRn file path and the name of DMR detection method.}

\item{threshold}{The methylation difference threshold t, default 0-0.9.}

\item{Title}{The title of rank graph, default NULL.}
}
\description{
This function will draw a line chart for DMRn.
}
\examples{
library("ggplot2")
DMRn_file = read.table(system.file('extdata','rank_DMRn.txt',package='DMRArrayMetric'))
for(i in 1:7){
   DMRn_file[i,1] = eval(parse(text=DMRn_file[i,1]))

}
rank_graph(DMRn_file,threshold=seq(0,0.9,0.1),Title='GroupA-GroupB')
}
